improve delete

+++ src/routes/dashboard/categories/category-form/service.ts
@@ -1,7 +1,9 @@
 import { tailwindColors, types } from '$lib/category/utils';
 import type { Category } from '$lib/category/utils';
+import { createModal } from '$lib/components/dialog/service';
 import { db } from '$lib/firebase';
 import { addDoc, collection, deleteDoc, doc, updateDoc } from 'firebase/firestore';
+import { EventDispatcher } from 'svelte';
 import { string } from 'yup';
 
 export function buildEmptyCategory() {
@@ -23,9 +25,15 @@ export function addCategory(data: Omit<Category, 'id'>, userId: string) {
 	return addDoc(categoriesCollectionRef, data);
 }
 
-export async function deleteCategory(id: string | undefined, userId: string) {
-	if (id) {
+export async function deleteCategory(
+	id: string | undefined,
+	userId: string,
+	dispatch: EventDispatcher<{ close: null }>,
+) {
+	const result = await createModal({ title: 'Are you sure?' });
+	if (result && id) {
 		const categoryDocRef = doc(db, 'users', userId, 'categories', id);
 		await deleteDoc(categoryDocRef);
+		dispatch('close');
 	}
 }
